+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Dependencies: bump textual from 0.85.0 to 0.85.1 #35
+
 ### Removed
 
 ## [0.4.1] - 2024-10-27 - Chasing Two Rabbits