+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add more torrent list sort orders [#38](https://github.com/anlar/tewi/issues/38)
 - Add torrent pieces info to torrent details
 - Add torrent from local file [#45](https://github.com/anlar/tewi/issues/45)
+- Show number of peers in torrent item [#47](https://github.com/anlar/tewi/issues/47)
 
 ### Changed